GradProject/requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.7
django-cors-middleware==1.3.1
django-rest-swagger==2.1.2
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
drf-writable-nested==0.3.2
idna==2.6
itypes==1.1.0
Jinja2==2.10
jsonfield==2.0.2
Markdown==2.6.11
markup==0.2
MarkupSafe==1.0
olefile==0.44
openapi-codec==1.3.2
Pillow==4.3.0
psycopg2==2.7.3.2
PyJWT==1.5.3
PyPDF2==1.26.0
pytz==2017.2
requests==2.18.4
requirements==0.1
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22

